backend/speckle_adapter/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
specklepy==2.17.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.3
